+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix pool destroying with tasks queuing enabled.
+
 ## [0.1.5] - 2023-12-18
 
 ### Added